backend/requirements.txt
fastapi
uvicorn
supabase
pandas
scikit-learn
python-dotenv
joblib
